+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Export
 Service to export YourLoops data in selected formats, either csv, json or xls.
 
+## Unreleased
+### Changed
+- YLP-536 Listen to SIGTERM & SIGINT signals & quickly shutdown when there is no export in progress.
+
 ## 1.5.1 - 2020-10-29
 ### Engineering
 - YLP-257 Review SOUP doc destination